Add configuration files and update tests to track created car IDs

+++ src/main/java/ui/steps/CarSteps.java
@@ -19,14 +19,14 @@ public class CarSteps {
     }
 
     @Step("Покупка автомобиля пользователем")
-    public void buyNewCar(Long userID, Long carID) {
+    public void buyNewCar(int userID, int carID) {
         buyOrSaleCarPage.openPage()
                 .isPageOpened()
                 .setData(userID, carID);
     }
 
     @Step("Продажа автомобиля пользователем")
-    public void sellNewCar(Long userID, Long carID) {
+    public void sellNewCar(int userID, int carID) {
         buyOrSaleCarPage.openPage()
                 .isPageOpened()
                 .setData(userID, carID);
@@ -38,9 +38,9 @@ public class CarSteps {
     }
 
     @Step("Проверка успешности создания автомобиля и получение его id")
-    public Long checkCreateCarAndGetId() {
+    public int checkCreateCarAndGetId() {
         SoftAssertions softly = new SoftAssertions();
-        Long carId = createCarPage.getCarId();
+        int carId = createCarPage.getCarId();
         softly.assertThat(createCarPage.getStatusMessage()).contains("Successfully pushed");
         softly.assertThat(carId).isPositive();
         softly.assertThat(createCarPage.getStatusCode()).isEqualTo(201);
@@ -49,7 +49,7 @@ public class CarSteps {
     }
 
     @Step("Проверка что автомобиль куплен пользователем")
-    public boolean isCarBought(Long userID, Long carID) {
+    public boolean isCarBought(int userID, int carID) {
         return checkStatusCode() == 200;
     }
